# toy-note/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple toy-note API",
        "title": "Toy-note API",
        "contact": {
            "name": "Jacob Bishop",
            "url": "https://github.com/Jacobbishopxy",
            "email": "jacobbishopxy@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/delete-post/{id}": {
            "delete": {
                "description": "delete a post by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "delete a post by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.successMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.errorMessage"
                        }
                    }
                }
            }
        },
        "/delete-tag/{id}": {
            "delete": {
                "description": "delete a tag by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tag"
                ],
                "summary": "delete a tag by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "tag ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.successMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.errorMessage"
                        }
                    }
                }
            }
        },
        "/download-file": {
            "get": {
                "description": "download an affiliate by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "affiliate"
                ],
                "summary": "download an affiliate by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "affiliate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.downloadSuccess"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.errorMessage"
                        }
                    }
                }
            }
        },
        "/get-posts": {
            "get": {
                "description": "get all posts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "get all posts",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Post"
                            }
                        }
                    }
                }
            }
        },
        "/get-tags": {
            "get": {
                "description": "get all tags",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tag"
                ],
                "summary": "get all tags",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Tag"
                            }
                        }
                    }
                }
            }
        },
        "/save-post": {
            "post": {
                "description": "Save post can be used to create a new post or update an existing post.\nIf id is not provided, it will create a new post; Otherwise, it will update an existing post.\nForm-data should also be provided if the post has any new affiliate.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "create/update a post",
                "parameters": [
                    {
                        "description": "post data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Post"
                        }
                    },
                    {
                        "type": "file",
                        "description": "affiliate files",
                        "name": "files",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Post"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.errorMessage"
                        }
                    }
                }
            }
        },
        "/save-tag": {
            "post": {
                "description": "create a new tag or update an existing tag, based on whether the tag ID is provided",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tag"
                ],
                "summary": "create/update a tag",
                "parameters": [
                    {
                        "description": "tag data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Tag"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Tag"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.errorMessage"
                        }
                    }
                }
            }
        },
        "/search-posts-by-tags": {
            "get": {
                "description": "get posts by tags",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "get posts by tags",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "tag ids",
                        "name": "ids",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Post"
                            }
                        }
                    }
                }
            }
        },
        "/search-posts-by-title": {
            "get": {
                "description": "get posts by title",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "get posts by title",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "post title",
                        "name": "title",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Post"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.downloadSuccess": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "controller.errorMessage": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "controller.successMessage": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "string"
                }
            }
        },
        "entity.Affiliate": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "object_id": {
                    "type": "string"
                },
                "post_refer": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entity.Post": {
            "type": "object",
            "properties": {
                "affiliates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Affiliate"
                    }
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "subtitle": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Tag"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entity.Tag": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Post"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}